Repackage to a domain we actually own.
