--- src/main/java/me/ryder/savagecore/commands/Executable.java
@@ -1,12 +1,0 @@
-package me.ryder.savagecore.commands;
-
-import org.bukkit.command.CommandSender;
-
-public interface Executable {
-
-    boolean execute(CommandSender paramCommandSender, String[] paramArrayOfString);
-
-    String getDescription();
-
-    String getPermission();
-}
