--- src/main/java/me/ryder/savagecore/utils/Methods.java
@@ -1,21 +1,0 @@
-package me.ryder.savagecore.utils;
-
-import me.ryder.savagecore.persist.Conf;
-import me.ryder.savagecore.utils.FileManager.Files;
-import org.bukkit.ChatColor;
-
-public class Methods {
-
-    public static String pl(String msg) {
-        return ChatColor.translateAlternateColorCodes('&', msg);
-    }
-
-    public static String getPrefix(String msg) {
-        return pl(Files.messages.getFile().getString("Messages.Prefix") + msg);
-    }
-
-    public static void reloadConfiguration() {
-        Conf.load();
-        Files.messages.reloadFile();
-    }
-}
