+++ src/main/java/net/savagellc/savagecore/commands/Executable.java
@@ -1,0 +1,12 @@
+package net.savagellc.savagecore.commands;
+
+import org.bukkit.command.CommandSender;
+
+public interface Executable {
+
+    boolean execute(CommandSender paramCommandSender, String[] paramArrayOfString);
+
+    String getDescription();
+
+    String getPermission();
+}
