+++ src/main/java/net/savagellc/savagecore/events/DenyBlazeDrowning.java
@@ -1,0 +1,23 @@
+package net.savagellc.savagecore.events;
+
+import net.savagellc.savagecore.persist.Conf;
+import org.bukkit.entity.Blaze;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.Listener;
+import org.bukkit.event.entity.EntityDamageEvent;
+import org.bukkit.event.entity.EntityDamageEvent.*;
+
+public class DenyBlazeDrowning implements Listener {
+
+    @EventHandler
+    public void onBlazeDrown(EntityDamageEvent e) {
+        if (Conf.antiBlazeDrownToggle) {
+            if (e.isCancelled()) {
+                return;
+            }
+            if (e.getEntity() instanceof Blaze && e.getCause() == DamageCause.DROWNING) {
+                e.setCancelled(true);
+            }
+        }
+    }
+}
